+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Improvements
 
+- Expand draggable area in command bar.
 - [trivial] Omit commands in the Debug (ladybug) menu from the results of the Quick Action bar.